Fix module augmentation support for mocha
