--- types/mocha/lib/ms.d.ts
@@ -1,17 +1,0 @@
-export = milliseconds;
-
-/**
- * Parse the given `str` and return milliseconds.
- *
- * @see {@link https://mochajs.org/api/module-milliseconds.html}
- * @see {@link https://mochajs.org/api/module-milliseconds.html#~parse}
- */
-declare function milliseconds(val: string): number;
-
-/**
- * Format for `ms`.
- *
- * @see {@link https://mochajs.org/api/module-milliseconds.html}
- * @see {@link https://mochajs.org/api/module-milliseconds.html#~format}
- */
-declare function milliseconds(val: number): string;
