+++ types/mocha/tsconfig.json
@@ -19,8 +19,6 @@
     },
     "files": [
         "index.d.ts",
-        "lib/ms.d.ts",
-        "lib/interfaces/common.d.ts",
         "mocha-tests.ts"
     ]
 }
